Extract octocrab retry metrics into a dedicated submodule
Move `PodbotOctocrabRetryMetrics`, `RetryEventContext`, the
`record_octocrab_retry_event` recorder and `github_status_class`
classifier out of `github::mod` and into a new `github::retry_metrics`
submodule, along with the `retry_metrics_tests` unit suite. This
brings `github/mod.rs` back under the 400-line limit and colocates
the retry observability surface, its tests, and its callers.
